Update and rename Suspect.js to Vain.js
